+++ src/tests/Feature/AdminLoginTest.php
@@ -4,7 +4,6 @@ namespace Tests\Feature;
 
 use App\Models\Admin;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
 class AdminLoginTest extends TestCase
